requirements.txt
# Telegram Bot
python-telegram-bot==21.0.1

# AI
anthropic==0.18.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0

# GitHub Integration
PyGithub==2.2.0

# HTTP Requests
httpx==0.26.0
aiohttp==3.9.3

# Utilities
python-dotenv==1.0.0
tiktoken==0.6.0

# Web server (for webhook)
fastapi==0.109.2
uvicorn==0.27.1
